+++ kerchunk/tests/test_netcdf.py
@@ -1,5 +1,4 @@
 import os
-from typing import Any
 
 
 import fsspec
@@ -8,7 +7,7 @@ from packaging.version import Version
 import pytest
 from kerchunk import netCDF3
 
-import zarr
+from kerchunk.utils import refs_as_store
 
 xr = pytest.importorskip("xarray")
 
@@ -27,29 +26,19 @@ bdata = xr.Dataset({"data": data}, attrs={"attr0": 3}).to_netcdf(
 )
 
 
-def create_store(test_dict: dict, remote_options: Any = None):
-    if Version(zarr.__version__) < Version("3.0.0.a0"):
-        return fsspec.get_mapper(
-            "reference://", fo=test_dict, remote_protocol="s3", remote_options=remote_options
-        )
-    else:
-        fs = fsspec.implementations.reference.ReferenceFileSystem(fo=test_dict, remote_options=remote_options)
-        return zarr.storage.RemoteStore(fs, mode="r")
-
-
 def test_one(m):
     m.pipe("data.nc3", bdata)
     h = netCDF3.netcdf_recording_file("memory://data.nc3")
     out = h.translate()
 
-    store = create_store(out, remote_options={"remote_protocol": "memory"})
+    store = refs_as_store(out, remote_protocol="memory")
 
     ds = xr.open_dataset(
         store,
         engine="zarr",
         backend_kwargs={
             "consolidated": False,
-            "zarr_format": "2",
+            "zarr_format": 2,
         },
     )
     assert (ds.data == data).all()
